requirements-dev.txt
-e .
pytest
boilerplate
invoke
pytest_selenium
pytest_factoryboy
mistune
aloe_django
